making update for invalid marketresults
